backend/requirements.txt
sanic==24.6.0
sanic-ext==23.12.0
aiosqlite==0.20.0
python-dotenv==1.0.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
